+++ janitor-lang/src/main/java/com/eischet/janitor/env/JListClass.java
@@ -68,7 +68,7 @@ public class JListClass {
     public static JString __toJson(final JList self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
         try {
             arguments.require(0);
-            return process.getEnvironment().getBuiltinTypes().string(self.exportToJson(process.getRuntime().getEnvironment()));
+            return Janitor.string(self.exportToJson(process.getRuntime().getEnvironment()));
         } catch (JsonException e) {
             throw new JanitorNativeException(process, "error exporting json", e);
         }
@@ -82,13 +82,13 @@ public class JListClass {
                 ++count;
             }
         }
-        return process.getEnvironment().getBuiltinTypes().integer(count);
+        return Janitor.integer(count);
     }
 
     public static JList __filter(final JList self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
         final JanitorObject callable = arguments.getRequired(0, JanitorObject.class);
         if (callable instanceof JCallable func) {
-            JList result = process.getEnvironment().getBuiltinTypes().list();
+            JList result = Janitor.list();
             for (final JanitorObject e : self.janitorGetHostValue()) {
                 if (func.call(process, new JCallArgs("filter", process, Collections.singletonList(e))).janitorIsTrue()) {
                     result.add(e);
@@ -103,7 +103,7 @@ public class JListClass {
     public static JList __map(final JList self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
         final JanitorObject callable = arguments.getRequired(0, JanitorObject.class);
         if (callable instanceof JCallable func) {
-            final JList result = process.getEnvironment().getBuiltinTypes().list(self.janitorGetHostValue().size());
+            final JList result = Janitor.list(self.janitorGetHostValue().size());
             for (final JanitorObject e : self.janitorGetHostValue()) {
                 result.add(func.call(process, new JCallArgs("map", process, Collections.singletonList(e))));
             }
@@ -115,17 +115,17 @@ public class JListClass {
 
     public static JString __join(final JList self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
         final String separator = arguments.getOptionalStringValue(0, " ");
-        return process.getEnvironment().getBuiltinTypes().string(self.janitorGetHostValue().stream().map(JanitorObject::janitorToString).collect(Collectors.joining(separator)));
+        return Janitor.string(self.janitorGetHostValue().stream().map(JanitorObject::janitorToString).collect(Collectors.joining(separator)));
     }
 
     public static JList __toList(final JList self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
         arguments.require(0);
-        return process.getEnvironment().getBuiltinTypes().list(self.janitorGetHostValue());
+        return Janitor.list(self.janitorGetHostValue());
     }
 
     public static JSet __toSet(final JList self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
         arguments.require(0);
-        return process.getEnvironment().getBuiltinTypes().set(self.janitorGetHostValue().stream());
+        return Janitor.set(self.janitorGetHostValue().stream());
     }
 
     public static JBool __isEmpty(final JList self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
@@ -135,7 +135,7 @@ public class JListClass {
 
     public static JInt __size(final JList self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
         arguments.require(0);
-        return process.getEnvironment().getBuiltinTypes().integer(self.janitorGetHostValue().size());
+        return Janitor.integer(self.janitorGetHostValue().size());
     }
 
     public static JBool __contains(final JList self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
@@ -151,7 +151,7 @@ public class JListClass {
     public static JList __randomSublist(final JList self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
         final int count = arguments.getInt(0).getAsInt();
         if (count >= self.janitorGetHostValue().size()) {
-            return process.getEnvironment().getBuiltinTypes().list(self.janitorGetHostValue());
+            return Janitor.list(self.janitorGetHostValue());
         }
         final Random random = new Random();
         final Set<Integer> indexes = new HashSet<>();
@@ -160,7 +160,7 @@ public class JListClass {
         }
         final List<JanitorObject> result = new ArrayList<>(indexes.size());
         indexes.forEach(i -> result.add(self.janitorGetHostValue().get(i)));
-        return process.getEnvironment().getBuiltinTypes().list(result);
+        return Janitor.list(result);
     }
 
     public static JNull __put(final JList self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
@@ -198,11 +198,11 @@ public class JListClass {
         }
         if (arguments.size() == 2) {
             if (arguments.get(0) == JNull.NULL && arguments.get(1) == JNull.NULL) {
-                return self.getRange(process.getEnvironment().getBuiltinTypes().integer(0), process.getEnvironment().getBuiltinTypes().integer(self.size()));
+                return self.getRange(Janitor.integer(0), Janitor.integer(self.size()));
             } else if (arguments.get(0) == JNull.NULL) {
-                return self.getRange(process.getEnvironment().getBuiltinTypes().integer(0), arguments.getInt(1));
+                return self.getRange(Janitor.integer(0), arguments.getInt(1));
             } else if (arguments.get(1) == JNull.NULL) {
-                return self.getRange(arguments.getInt(0), process.getEnvironment().getBuiltinTypes().integer(self.size()));
+                return self.getRange(arguments.getInt(0), Janitor.integer(self.size()));
             }
             return self.getRange(arguments.getInt(0), arguments.getInt(1));
         }
@@ -223,11 +223,11 @@ public class JListClass {
         // LATER: make ranges assignable, too?
         if (arguments.size() == 2) {
             if (arguments.get(0) == JNull.NULL && arguments.get(1) == JNull.NULL) {
-                return self.getRange(process.getEnvironment().getBuiltinTypes().integer(0), process.getEnvironment().getBuiltinTypes().integer(self.size()));
+                return self.getRange(Janitor.integer(0), Janitor.integer(self.size()));
             } else if (arguments.get(0) == JNull.NULL) {
-                return self.getRange(process.getEnvironment().getBuiltinTypes().integer(0), arguments.getInt(1));
+                return self.getRange(Janitor.integer(0), arguments.getInt(1));
             } else if (arguments.get(1) == JNull.NULL) {
-                return self.getRange(arguments.getInt(0), process.getEnvironment().getBuiltinTypes().integer(self.size()));
+                return self.getRange(arguments.getInt(0), Janitor.integer(self.size()));
             }
             return self.getRange(arguments.getInt(0), arguments.getInt(1));
         }