+++ janitor-lang/src/main/java/com/eischet/janitor/env/JRegexClass.java
@@ -1,5 +1,6 @@
 package com.eischet.janitor.env;
 
+import com.eischet.janitor.api.Janitor;
 import com.eischet.janitor.api.JanitorScriptProcess;
 import com.eischet.janitor.api.types.functions.JCallArgs;
 import com.eischet.janitor.api.errors.runtime.JanitorRuntimeException;
@@ -19,7 +20,7 @@ public class JRegexClass {
         final JString str = arguments.getString(0);
         final Matcher matcher = self.janitorGetHostValue().matcher(str.janitorGetHostValue());
         if (matcher.find()) {
-            return process.getEnvironment().getBuiltinTypes().string(matcher.group(1));
+            return Janitor.string(matcher.group(1));
         } else {
             return JNull.NULL;
         }
@@ -35,11 +36,11 @@ public class JRegexClass {
 
     public static JanitorObject extractAll(final JanitorWrapper<Pattern> self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
         arguments.require(1);
-        final JList list = process.getEnvironment().getBuiltinTypes().list();
+        final JList list = Janitor.list();
         final JString str = arguments.getString(0);
         final Matcher matcher = self.janitorGetHostValue().matcher(str.janitorGetHostValue());
         while (matcher.find()) {
-            list.add(process.getEnvironment().getBuiltinTypes().string(matcher.group(1)));
+            list.add(Janitor.string(matcher.group(1)));
         }
         return list;
     }
@@ -49,7 +50,7 @@ public class JRegexClass {
         final JString string = arguments.getString(0);
         final JString with = arguments.getString(1);
         final Matcher matcher = self.janitorGetHostValue().matcher(string.janitorGetHostValue());
-        return process.getEnvironment().getBuiltinTypes().string(matcher.replaceFirst(with.janitorToString()));
+        return Janitor.string(matcher.replaceFirst(with.janitorToString()));
     }
 
     public static JanitorObject replaceAll(final JanitorWrapper<Pattern> self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
@@ -57,16 +58,16 @@ public class JRegexClass {
         final JString string = arguments.getString(0);
         final JString with = arguments.getString(1);
         final Matcher matcher = self.janitorGetHostValue().matcher(string.janitorGetHostValue());
-        return process.getEnvironment().getBuiltinTypes().string(matcher.replaceAll(with.janitorToString()));
+        return Janitor.string(matcher.replaceAll(with.janitorToString()));
     }
 
     public static JanitorObject split(final JanitorWrapper<Pattern> self, final JanitorScriptProcess process, final JCallArgs arguments) throws JanitorRuntimeException {
         arguments.require(1);
-        final JList list = process.getEnvironment().getBuiltinTypes().list();
+        final JList list = Janitor.list();
         final JString str = arguments.getString(0);
         final String[] parts = str.janitorGetHostValue().split(self.janitorGetHostValue().pattern());
         for (final String part : parts) {
-            list.add(process.getEnvironment().getBuiltinTypes().string(part));
+            list.add(Janitor.string(part));
         }
         return list;
     }