+++ janitor-lang/src/main/java/com/eischet/janitor/runtime/JanitorSemantics.java
@@ -91,9 +91,9 @@ public class JanitorSemantics {
      */
     public static @NotNull JanitorObject negate(JanitorScriptProcess process, final JanitorObject currentValue) throws JanitorRuntimeException {
         if (currentValue instanceof JInt v) {
-            return process.getEnvironment().getBuiltinTypes().integer(-v.getValue());
+            return Janitor.integer(-v.getValue());
         } else if (currentValue instanceof JFloat f) {
-            return process.getEnvironment().getBuiltinTypes().floatingPoint(-f.getValue());
+            return Janitor.floatingPoint(-f.getValue());
         } else {
             throw new JanitorNotImplementedException(process, "we can only negate numbers");
         }
@@ -131,13 +131,13 @@ public class JanitorSemantics {
     ) throws JanitorRuntimeException {
         try {
             if (leftValue instanceof JInt leftInteger && rightValue instanceof JInt rightInteger) {
-                return process.getEnvironment().getBuiltinTypes().integer(intOp.apply(process, leftInteger.getValue(), rightInteger.getValue()));
+                return Janitor.integer(intOp.apply(process, leftInteger.getValue(), rightInteger.getValue()));
             } else if (leftValue instanceof JFloat leftFloat && rightValue instanceof JFloat rightFloat) {
-                return process.getEnvironment().getBuiltinTypes().floatingPoint(floatOp.apply(process, leftFloat.getValue(), rightFloat.getValue()));
+                return Janitor.floatingPoint(floatOp.apply(process, leftFloat.getValue(), rightFloat.getValue()));
             } else if (leftValue instanceof JInt leftInteger && rightValue instanceof JFloat rightFloat) {
-                return process.getEnvironment().getBuiltinTypes().floatingPoint(floatOp.apply(process, leftInteger.getAsDouble(), rightFloat.getValue()));
+                return Janitor.floatingPoint(floatOp.apply(process, leftInteger.getAsDouble(), rightFloat.getValue()));
             } else if (leftValue instanceof JFloat leftFloat && rightValue instanceof JInt rightInteger) {
-                return process.getEnvironment().getBuiltinTypes().floatingPoint(floatOp.apply(process, leftFloat.getValue(), rightInteger.getAsDouble()));
+                return Janitor.floatingPoint(floatOp.apply(process, leftFloat.getValue(), rightInteger.getAsDouble()));
             } else if (dateOp != null && leftValue instanceof JDateTime leftDate && rightValue instanceof JDateTime rightDate) {
                 return dateTimeDateTimeOp.apply(process, leftDate, rightDate);
             } else if (dateOp != null && leftValue instanceof JDate leftDate && rightValue instanceof JDate rightDate) {
@@ -175,9 +175,9 @@ public class JanitorSemantics {
      */
     public static @NotNull JanitorObject multiply(JanitorScriptProcess process, final JanitorObject leftValue, final JanitorObject rightValue) throws JanitorRuntimeException {
         if (leftValue instanceof JString leftString && rightValue instanceof JInt rightInt) {
-            return process.getEnvironment().getBuiltinTypes().string(repeat(leftString.janitorGetHostValue(), rightInt.getValue()));
+            return Janitor.string(repeat(leftString.janitorGetHostValue(), rightInt.getValue()));
         } else if (leftValue instanceof JInt leftInt && rightValue instanceof JString rightString) {
-            return process.getEnvironment().getBuiltinTypes().string(repeat(rightString.janitorGetHostValue(), leftInt.getValue()));
+            return Janitor.string(repeat(rightString.janitorGetHostValue(), leftInt.getValue()));
         }
         return numericOperation(process,
                 "multiply",
@@ -260,8 +260,8 @@ public class JanitorSemantics {
                 (p, a, b) -> a - b,
                 JDuration::subtract,
                 JDuration::subtract,
-                (proc, jDate, jDate2) -> durationBetween(process.getEnvironment().getBuiltinTypes(), jDate, jDate2),
-                (proc, jDateTime, jDateTime2) -> durationBetween(process.getEnvironment().getBuiltinTypes(), jDateTime, jDateTime2),
+                (proc, jDate, jDate2) -> durationBetween(jDate, jDate2),
+                (proc, jDateTime, jDateTime2) -> durationBetween(jDateTime, jDateTime2),
                 JDuration::subtract,
                 null, false
         );
@@ -418,7 +418,7 @@ public class JanitorSemantics {
      */
     public static @NotNull JanitorObject add(JanitorScriptProcess process, final JanitorObject leftValue, final JanitorObject rightValue) throws JanitorRuntimeException {
         if (leftValue instanceof JString || rightValue instanceof JString) {
-            return process.getEnvironment().getBuiltinTypes().string(leftValue.janitorGetHostValue() + String.valueOf(rightValue.janitorGetHostValue()));
+            return Janitor.string(leftValue.janitorGetHostValue() + String.valueOf(rightValue.janitorGetHostValue()));
         }
         return numericOperation(process,
                 "add",
@@ -448,7 +448,7 @@ public class JanitorSemantics {
      */
     public static @NotNull JanitorObject increment(JanitorScriptProcess process, final JanitorObject currentValue) throws JanitorRuntimeException {
         if (currentValue instanceof JInt currentInteger) {
-            return process.getEnvironment().getBuiltinTypes().integer(currentInteger.getValue() + 1);
+            return Janitor.integer(currentInteger.getValue() + 1);
         } else {
             throw new JanitorNotImplementedException(process, "we can only increment numbers at the moment");
         }
@@ -464,7 +464,7 @@ public class JanitorSemantics {
      */
     public static @NotNull JanitorObject decrement(JanitorScriptProcess process, final JanitorObject currentValue) throws JanitorRuntimeException {
         if (currentValue instanceof JInt currentInteger) {
-            return process.getEnvironment().getBuiltinTypes().integer(currentInteger.getValue() - 1);
+            return Janitor.integer(currentInteger.getValue() - 1);
         } else {
             throw new JanitorNotImplementedException(process, "we can only decrement numbers at the moment");
         }
@@ -529,8 +529,8 @@ public class JanitorSemantics {
      * @param right the right date
      * @return the duration between the two dates
      */
-    public static JDuration durationBetween(final BuiltinTypes builtins, final JDate left, final JDate right) {
-        return builtins.duration(Duration.between(right.janitorGetHostValue().atStartOfDay(), left.janitorGetHostValue().atStartOfDay()).toDays(), JDuration.JDurationKind.DAYS);
+    public static JDuration durationBetween(final JDate left, final JDate right) {
+        return Janitor.duration(Duration.between(right.janitorGetHostValue().atStartOfDay(), left.janitorGetHostValue().atStartOfDay()).toDays(), JDuration.JDurationKind.DAYS);
     }
 
     /**
@@ -541,8 +541,8 @@ public class JanitorSemantics {
      * @param right the right date
      * @return the duration between the two dates
      */
-    public static JDuration durationBetween(final BuiltinTypes builtins, final JDateTime left, final JDateTime right) {
-        return builtins.duration(Duration.between(right.janitorGetHostValue(), left.janitorGetHostValue()).toSeconds(), JDuration.JDurationKind.SECONDS);
+    public static JDuration durationBetween(final JDateTime left, final JDateTime right) {
+        return Janitor.duration(Duration.between(right.janitorGetHostValue(), left.janitorGetHostValue()).toSeconds(), JDuration.JDurationKind.SECONDS);
     }
 
     /**