+++ janitor-lang/src/test/java/com/eischet/janitor/JanitorTest.java
@@ -47,7 +47,6 @@ public abstract class JanitorTest {
             public JanitorEnvironment getCurrentEnvironment() {
                 return TestEnv.env;
             }
-
         });
     }