+++ janitor-lang/src/test/java/com/eischet/janitor/types/JStringTestCase.java
@@ -1,6 +1,7 @@
 package com.eischet.janitor.types;
 
 import com.eischet.janitor.JanitorTest;
+import com.eischet.janitor.api.Janitor;
 import com.eischet.janitor.api.RunnableScript;
 import com.eischet.janitor.api.errors.compiler.JanitorCompilerException;
 import com.eischet.janitor.api.errors.runtime.JanitorRuntimeException;
@@ -51,13 +52,13 @@ public class JStringTestCase extends JanitorTest {
     @Test
     void basics() throws JanitorCompilerException, JanitorRuntimeException {
         OutputCatchingTestRuntime rt = OutputCatchingTestRuntime.fresh();
-        @NotNull JString string = rt.getEnvironment().getBuiltinTypes().string("foobar");
+        @NotNull JString string = Janitor.string("foobar");
         assertEquals("foobar", string.janitorGetHostValue());
         assertEquals("foobar", string.janitorToString());
         assertEquals("foobar", string.toString());
 
-        @NotNull JMap map = rt.getEnvironment().getBuiltinTypes().map();
-        @NotNull JInt oneTo = rt.getEnvironment().getBuiltinTypes().integer(1);
+        @NotNull JMap map = Janitor.map();
+        @NotNull JInt oneTo = Janitor.integer(1);
         map.put(string, oneTo);
         JanitorObject oneFro = map.get(string);
         assertEquals(oneFro, oneTo);